refactor: move project_with_review_dto to the google_sheet.dto package
